refactor(processing): split title_inserter along its two service seams
title_inserter.py was 954 lines doing three jobs, two of which are not
"insert titles into a clip list":

- TitleBackgroundRenderer (title_background_renderer.py) pre-renders the
  first and last clip through the assembly pipeline, so the title deblurs
  out of — and the ending rewinds into — the exact frames the assembler
  will produce. The two pre-renders built a byte-identical ffmpeg command
  twice; they now share one builder, same flags, same pixel formats, same
  colour tagging.
- TitleDividerPlanner (title_divider_planner.py) decides which divider
  cards a clip list gets and interleaves them. Every one of its methods
  took the same generator and the same title settings, so both are now
  injected once at construction behind a DividerCardGenerator protocol
  and each method only takes the clips it is dividing. Date parsing and
  month/year change detection use no state at all and moved out as
  module-level functions.

TitleInserter keeps the orchestration it is named for: resolve the
canvas, build the title config, generate the opening screen, delegate
dividers, generate the ending, pre-decide transitions, assemble.

Pure move. No colour constant, encoder flag, pixel format or filter
string changed, and the per-memory-type dispatch (month title, year
dividers, trip location cards) is untouched.

954 -> 434 lines, with 323 and 212 in the new modules.
