+++ complexipy-snapshot.json
@@ -3457,221 +3457,225 @@
       {
         "name": "TitleInserter::assemble_with_titles",
         "complexity": 22,
-        "line_start": 752,
-        "line_end": 954,
+        "line_start": 233,
+        "line_end": 434,
         "line_complexities": [
           {
-            "line": 770,
+            "line": 251,
             "complexity": 1
           },
           {
-            "line": 771,
+            "line": 252,
             "complexity": 0
           },
           {
-            "line": 773,
+            "line": 254,
             "complexity": 0
           },
           {
-            "line": 774,
+            "line": 255,
             "complexity": 2
           },
           {
-            "line": 775,
+            "line": 256,
             "complexity": 0
           },
           {
-            "line": 779,
+            "line": 260,
             "complexity": 1
           },
           {
-            "line": 781,
+            "line": 262,
             "complexity": 0
           },
           {
-            "line": 783,
+            "line": 264,
             "complexity": 0
           },
           {
-            "line": 784,
+            "line": 265,
             "complexity": 0
           },
           {
-            "line": 786,
+            "line": 267,
             "complexity": 0
           },
           {
-            "line": 789,
+            "line": 270,
             "complexity": 0
           },
           {
-            "line": 790,
+            "line": 271,
             "complexity": 1
           },
           {
-            "line": 791,
+            "line": 272,
             "complexity": 0
           },
           {
-            "line": 798,
+            "line": 279,
             "complexity": 0
           },
           {
-            "line": 808,
+            "line": 289,
             "complexity": 0
           },
           {
-            "line": 811,
+            "line": 292,
             "complexity": 1
           },
           {
-            "line": 814,
+            "line": 295,
             "complexity": 0
           },
           {
-            "line": 815,
+            "line": 296,
             "complexity": 0
           },
           {
-            "line": 818,
+            "line": 299,
             "complexity": 0
           },
           {
-            "line": 819,
+            "line": 300,
             "complexity": 2
           },
           {
-            "line": 820,
+            "line": 301,
             "complexity": 0
           },
           {
-            "line": 826,
+            "line": 307,
             "complexity": 0
           },
           {
-            "line": 828,
+            "line": 309,
             "complexity": 1
           },
           {
-            "line": 829,
+            "line": 310,
             "complexity": 2
           },
           {
-            "line": 830,
+            "line": 311,
             "complexity": 0
           },
           {
-            "line": 840,
+            "line": 321,
             "complexity": 2
           },
           {
-            "line": 843,
+            "line": 324,
             "complexity": 0
           },
           {
-            "line": 855,
+            "line": 336,
             "complexity": 0
           },
           {
-            "line": 869,
+            "line": 350,
             "complexity": 2
           },
           {
-            "line": 872,
+            "line": 353,
             "complexity": 0
           },
           {
-            "line": 873,
+            "line": 354,
             "complexity": 1
           },
           {
-            "line": 877,
+            "line": 358,
             "complexity": 0
           },
           {
-            "line": 885,
+            "line": 359,
             "complexity": 0
           },
           {
-            "line": 888,
+            "line": 365,
+            "complexity": 0
+          },
+          {
+            "line": 368,
             "complexity": 1
           },
           {
-            "line": 889,
+            "line": 369,
             "complexity": 2
           },
           {
-            "line": 904,
+            "line": 384,
             "complexity": 0
           },
           {
-            "line": 907,
+            "line": 387,
             "complexity": 1
           },
           {
-            "line": 914,
+            "line": 394,
             "complexity": 0
           },
           {
-            "line": 919,
+            "line": 399,
             "complexity": 0
           },
           {
-            "line": 925,
+            "line": 405,
             "complexity": 0
           },
           {
-            "line": 926,
+            "line": 406,
             "complexity": 0
           },
           {
-            "line": 927,
+            "line": 407,
             "complexity": 0
           },
           {
-            "line": 928,
+            "line": 408,
             "complexity": 0
           },
           {
-            "line": 933,
+            "line": 413,
             "complexity": 2
           },
           {
-            "line": 937,
+            "line": 417,
             "complexity": 0
           },
           {
-            "line": 938,
+            "line": 418,
             "complexity": 0
           },
           {
-            "line": 939,
+            "line": 419,
             "complexity": 0
           },
           {
-            "line": 940,
+            "line": 420,
             "complexity": 0
           },
           {
-            "line": 941,
+            "line": 421,
             "complexity": 0
           },
           {
-            "line": 942,
+            "line": 422,
             "complexity": 0
           },
           {
-            "line": 947,
+            "line": 427,
             "complexity": 0
           },
           {
-            "line": 949,
+            "line": 429,
             "complexity": 0
           }
         ]
       }
     ],
-    "complexity": 157
+    "complexity": 51
   },
   {
     "path": "src/immich_memories/security.py",
