+++ vulture-whitelist.py
@@ -215,8 +215,8 @@ has_any_hdr_clip  # unused function (src/immich_memories/processing/hdr_utilitie
 _.is_burst  # unused property (src/immich_memories/processing/live_photo_merger.py:46)
 mtime_ns  # unused variable (src/immich_memories/processing/probe_cache.py:24)
 _.invalidate  # unused method (src/immich_memories/processing/probe_cache.py:162)
-_.get_orientation_from_clips  # unused method (src/immich_memories/processing/title_inserter.py:405)
-_.get_resolution_tier  # unused method (src/immich_memories/processing/title_inserter.py:421)
+_.get_orientation_from_clips  # unused method (src/immich_memories/processing/title_inserter.py:185)
+_.get_resolution_tier  # unused method (src/immich_memories/processing/title_inserter.py:201)
 _.get_target_size  # unused method (src/immich_memories/processing/transforms.py:136)
 _._get_video_dimensions  # unused method (src/immich_memories/processing/transforms.py:201)
 _._transform_fit  # unused method (src/immich_memories/processing/transforms.py:205)
